+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [0.1.2] - 2023-05-18
+
+- The size of the Window should be set up in the config file
+- Default keybindings can be remapped 
+- Added scrollbars
+
 ## [0.1.1] - 2023-05-17
 
 - Readme update